// src/generated/model/createMappingRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CreateMappingRequestStatus } from './createMappingRequestStatus.ts';

export interface CreateMappingRequest {
  programId: string;
  programLevelId: string;
  feeTemplateId: string;
  effectiveDate: string;
  status: CreateMappingRequestStatus;
  remarks?: string;
}