// src/generated/model/updateMappingRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UpdateMappingRequestStatus } from './updateMappingRequestStatus.ts';

export interface UpdateMappingRequest {
  feeTemplateId: string;
  effectiveDate: string;
  status: UpdateMappingRequestStatus;
  remarks?: string;
}