// src/generated/model/mappingResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MappingResponseStatus } from './mappingResponseStatus.ts';

export interface MappingResponse {
  id?: string;
  programId?: string;
  programLevelId?: string;
  feeTemplateId?: string;
  effectiveDate?: string;
  status?: MappingResponseStatus;
  remarks?: string;
  createdAt?: string;
  updatedAt?: string;
}